pyproject.toml
[tool.poetry]
name = "rio-test"
version = "0.1.0"
description = ""
authors = ["Jean-Pierre Chauvel <jean.p.chauvel@gmail.com>"]
license = "CC0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
rio-ui = {extras = ["window"], version = "^0.6.4"}

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
